pyproject.toml
[project]
name = "binny-buddy-server"
version = "0.1.0"
description = ""
authors = [
    {name = "tonynamy",email = "me@tonynamy.dev"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.0,<5.3.0)",
    "psycopg2-binary (>=2.9.10,<2.10.0)",
    "backoff (>=2.2.1,<2.3.0)",
    "gunicorn (>=23.0.0,<23.1.0)",
    "django-ninja (==1.4.1)",
    "requests (==2.32.3)",
    "mysqlclient (==2.2.4)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "~0.11.5"